requirements.txt
absl-py==2.1.0
ansicon==1.89.0
astunparse==1.6.3
attrs==24.2.0
autobahn==19.11.2
Automat==24.8.1
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
coverage==7.6.4
cryptography==43.0.3
editor==1.6.6
flake8==7.1.1
Flask==3.0.3
flatbuffers==24.3.25
gast==0.6.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio>=1.48.0,<2.0.0
grpcio-status>=1.48.0,<2.0.0
gTTS==2.5.3
h5py==3.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
inquirer==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jinxed==1.3.0
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.1
mock==5.1.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf>=3.20.0,<5.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pyttsx3==2.98
pyuseragents==1.0.5
PyYAML==6.0.2
readchar==4.2.0
requests==2.32.3
responses==0.25.3
rich==13.8.1
rsa==4.9
runs==1.2.2
safeIO==1.2
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0  # Use standard tensorflow
termcolor==2.4.0
translatepy==2.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xmod==1.8.1
zope.interface==7.1.1
